Split GreedyTraversalPlan
